requirements.txt
scikit-learn
numpy
scipy
tensorflow
torch
torchvision

absl-py
appnope
astor
attrs
backcall
bleach
cachetools
certifi
chardet
ConfigArgParse
cycler
decorator
defusedxml
entrypoints
future
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
jedi
Jinja2
joblib
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
Keras-Applications
Keras-Preprocessing
kiwisolver
Markdown
MarkupSafe
matplotlib
mistune
more-itertools
nbconvert
nbformat
notebook

oauthlib
opt-einsum
pandas
pandocfilters
parso
pexpect
pickleshare
Pillow
pluggy
prometheus-client
prompt-toolkit
protobuf
ptyprocess
pyasn1
pyasn1-modules
Pygments
pyparsing
pyrsistent
python-dateutil
python-jsonrpc-server
python-language-server
pytz
PyYAML
pyzmq
qtconsole
requests
requests-oauthlib
rsa
Send2Trash
six
scikit-learn
tabulate
tensorboard
tensorflow-estimator
termcolor
terminado
testpath
tornado
traitlets
urllib3
virtualenv
virtualenv-clone
wcwidth
webencodings
Werkzeug
widgetsnbextension
wrapt
zipp
